pyproject.toml
[project]
name = "rlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.2.0",
    "langgraph>=0.0.45",
    "pydantic>=2.0",
    "numpy>=1.24",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "mistralai>=0.4.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
